Jetifier: Auto-fallback for inner types.
Types map not longer works with nested types (A$B$1). But instead it
only maps the root types (A). Any inner type is then automatically
mapped to the same place as its parent.

By this we are loosing functionality to move inner types somewhere else
but given we never used that it's cool.

Test: Added
Bug: b/74446201

Change-Id: I06d27dab3f50c42c093eb56e80b9dfb8507af52e