[WebView Support Library] Suppress flaky WebViewCompat.testCheckThread
testCheckThread calls an API on the wrong thread and ensures it throws a
certain exception. But currently, when running on devices with WebView
APKs that are incompatible with the webview support library, we throw a
different exception related to the WebView APK code reflection failing.
Disable this test temporarily until we can fix this properly in
b/76202025.

Bug: 76172292
Test: run androidx.webkit tests.
Change-Id: If1ef44e438dd9873d1083dfe1275469f239f7dea