Reduce the font size of "Advanced" option
Advanced option is created by attribute ：
"initialExpandedChildrenCount" and the corresponding layout is
expand_button.xml. Reduce the title of "Advanced" option to 14sp, by
modify the attribute of expand_button's title. Set textAppearance to
textAppearanceSmall to reduce the size, and override the textcolor
to textColorPrimary to let text color remain the same.

Bug: 72953904
Test: None
Change-Id: I03c98788da4433af7818f6238dcbe74d7fdb2143