Only match addresses without zip codes if at the end.
The old native implementation of findAddress only accepted addresses
without zip codes if they appeared at the end of the string. This was
probably a bug as the documentation implies this should work in all
cases, but fixing this bug has caused a lot of false positives while not
fixing very many false negatives and this functionality is being
deprecated anyway, so change it back.

Remove LinkifyCompatTest#testFindAddress_withoutZipCode because the
LinkifyCompatTest mechanisms expect to be able to add surrounding text
and still have the address match, which is no longer the case with this
change. The coverage of this isn't vital.

Fixes: 77638722
Test: FindAddressTest & LinkifyCompatTest
Change-Id: Iff713173444fc3318cdcfb679022233a6a177720