MediaSessionCompat: Fix setPlaybackToLocal(stream) in API < 21
In API < 21, MediaSessionCompat.setPlaybackToLocal(int stream) did not
set the given stream to the session. As a result, the session could
not control any volume except the default stream (STREAM_MUSIC).
This CL fixes this behavior so that the sesson can control volume of
other streams (e.g. STREAM_ALARM).

This CL also fixes the failures of below tests in API < 21:
 - MediaController2Test.testSetVolumeWithLocalVolume
 - MediaController2Test.testAdjustVolumeWithLocalVolume

Bug: 78155309
Test: Ran version-compat-test
Change-Id: I1122cf2c917653b5baed1c1d2f4b604c01913369