Switch WorkRequest/WorkStatus id's to UUIDs.
Per API council feedback.

I chose not to change the internal WorkSpec id type
because that would lead to a lot of TypeConversions
that really aren't needed at runtime (as most of our
queries deal with id's).

To make testing a little easier, I included a new
method in WorkRequest called getStringId.

Change-Id: If9e5d7409834465704ac28cd7f1589425e936544
Fixes: 78288868
Test: Ran tests.