Convert MessagingStyle constructor to use Person.
Replaces the CharSequence constructor with a Person constructor and
removes the mUserDisplayName with just mUser instead. This allows the
benefits of Person (ie. adding an avatar) within MessagingStyle. This CL
also fixes rendering of compat-built MessagingStyle notifications. It
seems like frameworks depends on an extra that was previously set which
was no longer set by an update. The following are screenshots on various
configurations.

API 23:
 no title, group: https://screenshot.googleplex.com/cbJSN3h5HxW.png
 no title, no group: https://screenshot.googleplex.com/5LHHss1GPhB.png
 title, group: https://screenshot.googleplex.com/PXPpcd8jFi2.png
 title, no group: https://screenshot.googleplex.com/u2XuOQfyn2w.png

API 27:
 no title, group: https://screenshot.googleplex.com/RNNzy7Ei6v7.png
 no title, no group: https://screenshot.googleplex.com/VpCq7yJMQ9x.png
 title, group: https://screenshot.googleplex.com/VWjxV6o2ht0.png
 title, no group: https://screenshot.googleplex.com/KAp6QL3nLhe.png

On another note. This CL cleans up a bunch of deprecated calls to
Message#getSenderName for #getPerson.

Also cleans up some scope issues for fields and methods.

Test: ./gradlew core:connectedCheck
Change-Id: I3ef2519e297664d170383e35e0bef722246c4312
Fixes: 78310130
Fixes: 78598160