Fix conversation titles being rendered in SDK < P.
Conversation titles shouldn't be rendered when a notification isn't a
group in SDK < P. For some reason, compat uses the same extras key as
core framework when passing notification title data through, and in
doing so, will incorrectly set conversation titles in SDK < P when there
is no group conversation.

This CL also adds a toggle-able test to preview MessagingStyle
notifications by simply enabling a test and running on a device or
emulator.

Test: ./gradlew core:connectedCheck
Change-Id: I34367281e9445e6e06b2abcdfbf531cbca708271
Fixes: 79432786