Update compat RemoteInput#setChoices documentation
For apps that target < P, RemoteInput#setChoices doesn't render on
mobile devices. When apps target >= P, these choices will become visible
in the notification shade. This behavior was previously undocumented,
but due to this change, documentation must be updated to inform
consumers of intent and workarounds.

This CL also update a bunch of the builder docs and improves kotlin
interop with nullable/nonnull annotations.

Test: none
Issue: 79244189
Change-Id: I749aee13df6d3b0dd94e1525d4a83d2997dbe04b