Add special behavior for SubheaderListItems.
The SubheaderListItem should hide the divider for itself and the item
before it. Rather than having special logic in each ListItem to handle
this, make it a special case within the ListItemAdapter.

The adapter will check for the existence of subheaders and hide dividers
before it. Also hardcode the hidding of subheaders' dividers.

Fixes: 79748701
Test: ./gradlew car:connectedCheck
Test: Boot up list with subheaders on sample app and verify that
subheaders before and after are hidden

Change-Id: Ife9d261df948b04463e3944f7f230d52a200b8b1