Fixes issues when switching between small / large and scroll / no scroll
Non-scrolling slices weren’t able to update between small and large
because the height being passed to updateDisplayedItems was the
current height (i.e. small) and then it would restrict the items to
that height; this fixes that by passing default large height.

This uncovered an issue where the height for large non-scrolling slices
wasn’t stable; this is fixed by some reworking in onMeasure and removing
maxLarge height / “ideal item count” for non-scrolling slices which was
dumb.

Also fixes issue where SliceView didn’t handle invalid modes well.

Also breaks up the SliceView#reinflate method so that only the relevant
parts of it are called when necessary.

Test: ./gradlew slice-view:connectedCheck, look at screenshots
      Manual - test wifi and grocery slice, enable / disable scrolling
               and switching between modes
             - sanity check: test reservation and ride slice to ensure
               they’re not clipped weirdly in non-scrolling mode
Fixes: 79785423
Fixes: 78912603
Bug: 79960847
Change-Id: I6cf53e1e8fd93bfa80437507908d7eec84e465fd