Bring consistency to top margin for the scroll bar.
The PagedScrollBarView allows its top margin to be customized. However,
it was including a top padding with the same default value. This meant
that applications setting the top margin to 0 would still see the top
padding.

Refactor so that there is only margin affecting the top of the
PagedScrollBarView.

Fixes: 80433726
Test: build sample application and verify that top margin is the same as
before and can be customized.
Test: ./gradlew car:connectedCheck

Change-Id: I594bc9c0f94ab4e56d431ca610665676fbc4f8db