Links visibility of a PreferenceGroup to its children
Hiding a PreferenceGroup will now also hide its children. To see if a
preference is actually shown to the user, use Preference#isShown(). This
behaviour makes more sense logically and is also similar to how Views
work in a hierarchy with getVisibility() and isShown().

Bug: b/79926442
Test: Added
Change-Id: I109d9ba0e7e06ae16dfce5d8c7687d9f2ad03d4b
Merged-In: I140ba92accf6185403b527bcfb5b8496bcd62aba