Deprecate NavOptions' clear task
As popUpTo the root of your graph with
inclusive = true is equivalent to clear
task, there's no need for a separate flag.

This reduces confusion where developers could
set clear task to be true and a popUpTo
destination and be ensure which would take
precedence. It also fixes issues with
Activity destinations that would not interact
well with clear task.

Test: tests still pass
BUG: 80338878
Change-Id: I98192d1b41cfa8e44018123957eea842f00fc6f7