MediaSessionCompat: Allow removing callback in no-looper thread
The implementation of MediaSessionCompat#setCallback(Callback, Handler)
creates a new Handler when the given handler is null.

This behavior is mostly right. However, when the given callback is null
(i.e. removing callback from the session), creating a Handler should not
be needed.

This CL makes MediaSessionCompat#setCallback create a Handler only when
the given callback is not null, so that setCallback(null) can be called
in thread with no looper.

Test: ./runtest.sh 4 (from API 19, 21 to P devices)
Change-Id: I74d9446232e5576ba88cb404d9b1758e64028da4