Fix overlapping Fragments caused by single top
FragmentNavigator handles navigate calls with the
single top flag by replacing the existing Fragment.

However, when the Fragment is on the FragmentManager
back stack, a replace() operation on its own will
visually replace the Fragment, but after hitting the
back button the replaced Fragment will still be visible,
overlapping the previous Fragment on the back stack.

Instead, a single top operation should pop the Fragment
off the FragmentManager back stack and put the new
Fragment in its place.

Test: new FragmentNavigatorTest tests pass
BUG: 79407969
Change-Id: I15c07f7bf4e92f9d036f202742d985416b2007bb