[WebView Support Library] move workaround logic into test
This moves workaround logic for old-style versionNames in Webview
packages. The old format was to specify only the major version number,
followed by miscellaneous text wrapped in parentheses.

This moves the logic (which turns e.g. "38 (...)" into "38.0.0.0") out
of the WebViewVersion class and into the test case itself, and adds a
comment to the WebViewVersion class pointing to documentation on
Chromium version numbers.

This does not change testWebViewVersionMatchesInstrumentationArgs
because that only checks string equality. If the versionName follows the
pre-M42 format, it will fail the test as expected.

No expected change to behavior.

Bug: 110278473
Test: ./gradlew :webkit:connectedAndroidTest (on L device)
Change-Id: Ice29a20a00729599d426a500476fa94ab3a3fa42