MCV2: Store seek commands as a queue
Currently, seekTo(long) commands are sent to MediaPlayer2 every time
the progress bar is scrolled. As a result, when scrolling
through a file, there is significant lag after one stops scrolling.
In order to prevent the lag, this CL implements a queue for holding
the commands and removing the unnecessary ones.

Bug: 109713730
Test: Manually
Change-Id: I841aa570830d57a3e6e84455e3a45dfa8ab1bb34