Avoid popping the whole back stack from invalid IDs
If popBackStack is called with a destination ID that
doesn't exist on the back stack, popBackStack shouldn't
take that as an opportunity to pop the entire back stack.

Instead, we should log and return false, allowing developers
to explicitly look for this signal if they care.

Treat popping on an empty state similarly - just return false.

Test: new NavControllerTest
BUG: 110893637
Change-Id: I9740676cdae0f91616ae974710c79820c8db82be