Fix setting a specific style on a slice
This was working if you set it in the theme only.

This is because the style on the AttributeSet gets
cleared after the view is done inflating?? This is
based on my debugging. This is fixed by extracting
all the attributes we care about during view
construction rather than caching the AttributeSet
and looking things up from it later.

Test: ./gradlew slice-view:connectedCheck
      manual - set style specifically in XML on SliceView
             - set style via theme
Fixes: 77340278
Change-Id: I028a50fcf5787277c7e3c33bbacd56163b69b539