Skip AppCompatTextViewTest#testTextSize_canBeZero
The test was introduced in I691c1ff6521a4eaed07735f6e51c532c6ba46fd7
(see that for background), and before this CL it was failing on sdk
15. The problem seems to be that, on sdk <= 15, it used to be impossible
to setTextSize to 0 (setting it programatically did not work either, in
contrast to sdk 28), and the origin of the problem seems to be in the
native layer. This CL skips the test for sdk <= 15.

Test: ./gradlew appcompat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.appcompat.widget.AppCompatTextViewTest
Bug: 111322333
Change-Id: Ie48ceeaf24e9ad73a0b91848c79b414c13cd5a19