Add accessibilityPane to viewCompat
Add  the pane title property like every other support property
Lots to this, notifyViewAccessibilityStateChangedIfNeeded
behaves the same way as it does in the framework...
The pane appeared notifications are sent (adding disappeared constant
but not sending it, so service developers can prep for it, but it won't
send annoying events to services that arent ready for it yet).
Take note of AccessibilityPaneVisibilityManager.
First of all we needed a way to determine when visibility has changed,
so I uses ViewTreeObserver.addOnGlobalLayoutListener.
But, the VTO is disassociated from the view when
the view is detached from the window. So onViewAttachedToWindow
reregister it.
In the implementation of view there is a concept of
mFloatingTreeObserver, where you can create a VTO when the view isn't
attached to any window. But I still check in addAccessibilityPane,
because you can register the same object twice, and probably get
multiple callbacks. I could also get rid of a line of code by putting
the reregister in viewDetached callback, but that seems a bit confusing
and easy to break.

Test: Added a few unit tests.
Bug:80099879
Change-Id: I16a3b3afee8fb05174cf48050a80bf78b2593e40