Create unique WorkRequests when using WR.Builder#build()
Previously, we created the id and WorkSpec in the Builder
constructor, meaning subsequent calls to Builder#build()
would yield the same object.  This is really not how
Builders are supposed to work.

Fixing this required making copy constructors for several
internal classes.

Added some finals to methods that shouldn't be overriden.

Change-Id: Ieecbb0ca7d755e7ec65699fc77e589757f14a9d0
Fixes: 111408337
Test: Added and ran tests in WorkTest.