Handle false-negatives for force stop status.
The force stop check relies on a known
PendingIntent no longer existing on startup.
This works for the vast majority of cases,
but we have been able to reproduce a case on
older devices where the PendingIntent still
exists upon an immediate app restart.  This
PendingIntent just hasn't been GC'ed yet.

This is a very rare case that's mostly only
reproducible if you're trying to force this
situation on older devices (perhaps ART helps
here?).  Since there is really no good
solution here, we've opted for the option of
doing 3 backed-off retries to make sure we
catch the force stop on process start.  In
practice, the PendingIntent is cleaned up in
just a few seconds, but since we are dealing
with the unpredictability of GC's, it's
better to be overly-cautious.

Fixes: 111765866
Test: Updated and ran tests
Change-Id: Id804464d6ea8b7a8ed5ba5256c6303fde6dad929