Revert "Handle false-negatives for force stop status."
This reverts commit 2dda2580f1506eacd0be0d5273b65823deb3a031.

Reason for revert: This change was submitted prematurely.  This isn't actually solving the issue; it appears that the PendingIntent exists if the app is relaunched right away even if you wait a while before checking.  We don't have any great workarounds, and this is definitely a very rare edge case, so we will not be fixing this.

Change-Id: Iafa90ac516711ebc3b435702379f017ab9fad158