Fix dividers not drawing in the right place during refreshes.
During a refresh, an item can be pushed off-screen but still exist
during the iteration through each child. As a result, that item's top
value will not match its assumed position, resulting in a "spacing"
value is is negative.

Do not draw a divider in these cases to avoid the divider appearing in
the wrong place.

Fixes: 112053354
Test: Add a TextListItem that refreshes and verify that there is no
flash of dividers after a refresh of the list.

Change-Id: Ib24ab30d8b99b72575512c1f6510659b8a03549f