Add a few bindings.dart
The CL adds:
- widgets/bindings.dart
- foundation/bindings.dart
- window.dart
- other stubs to keep the project green
Note that certain things I left commented are likely to be required for
"Hello world", but I decided to stop before the number of stubs required
explodes in size again.

The classes I added were making frequent use of Dart Futures. I replaced
this with Kotlin Deferred, which solves the problem but is not really
idiomatic.

Test: built it
Change-Id: Id04a71aa0173c866e616720cb3923a25ee1d7218