MCV2: Fix inconsistent transport controls location
When rotating the screen, the transport controls location alternates
between the center view and the bottom bar view.

This CL fixes this behavior by removing the code for checking if the
current transport controls width is 0. Previously, since the
mTransportControls variable was being inflated whenever the mSizeType
was changed, the current transport controls width was 0, preventing
the layout from being properly updated.

Bug: 112523678, Bug: 111246858
Test: Manually
Change-Id: I9eeb2a76dd2d62db0585c025cd6e4f1931d62aa9