Migrate some image stuff.
This is more work toward the hello world case. The most primitive item
I found that we would run through the flutter logic is an image. Text is
obviously very hard. So I'm migrating stuff that we need to show an image
and run some hello world app for it.

I will actually try later to swap Android's bitmap there and render
something. But first step will be just debug output in the paint method.

Test: TBD

Change-Id: I59e953d99fda6afe71c185f5a2711a633be886f0