Port SceneBuilder and related native code. Simple image drawing hello world
As a result we have a simple hello world which is really drawing the image on a Android custom view's canvas!
I added SimpleFlutterView for it and now in SimpleFlutterActivity we can just call setContentView(SimpleFlutterView(this, widget)).
SimpleFlutterView is registering itself as a delegate for a Window.Render method and then we are drawing the Scene on a Canvas.

Comparing to the prev. version of CL I simplified the code a bit by commenting out more code we don't currently gonna use, removed all Skia wrappers and using our Canvas, Picture, Rect and only left one class SkMatrix, as I need a wrapper on a Framework's Matrix, its hard to reimplement the same with our Matrix3, we would need to port methods like concat, which we have in a Framework's Matrix.
Also ported Picture and PictureRecorder a bit and mocked some methods in semantics to not block hello world.

About poring of the methods we currently use from SceneBuilder:
Currently we have ported two layers which are added into the scene:
1) TransformLayer created in RenderView - root layer for any Flutter View
2) PictureLayer created in PaintingContext - everything we gonna drow on canvas will be recorded into this picture
SceneBuilder then passing this two layers into the native c++ code from 'flow' package, where there is yet another set of abstractions for layers with
the same names, but the bit different logic. You can think about this layers as Android's view hierarchy. There are containers like TransformLayer and simple nodes like PictureLayer. And this layers has two main methods:
1) Preroll - for like a measuring (calculate bounds)
2) Paint - draw on a canvas
Then we have a CompositorContext which has a method AcquireFrame where we pass a canvas to get a Frame object, then we build a Scene with SceneBuilder and are able to perform drawing:
val frame = CompositorContext().AcquireFrame(canvas)
val layerTree = scene.takeLayerTree()
layerTree.Preroll(frame)
layerTree.Paint(frame)

Test: n/a
Change-Id: Ifb9ba4c70a3282f59a43299e2589377f4bd65101