Fix flaky MediaSession2_KeyEventTest methods
1) For testPause/testPlayPause_pause:
The tests flaked due to an unclosed session in MediaSession2Test.
During the failing media key event tests, media button session was
changed to that unclosed session, so our target session couldn't get
any media key events.

2) For testStop:
This CL doesn't include the fix for failure of testStop, but is removing
@FlakyTest annotation since it did not flake.

Also, this CL renames the names of test methods with more clear names.

Bug: 115432583, Bug: 115464664, Bug: 112291143
Test: Ran all tests in service module 20 times.
      Without this fix, the test flakes.
Change-Id: I0572ccf5b6098ec6cb6630d8666fe923ebab6aa7