Introduce thread assertions for BindingBase Deferred usage.
We discussed with Flutter team and it turned out they are not using Deferred for multithreading at all. It is more like a callback to execute later on the same thread. But with Kotlin's coroutines it is extreamly easy to launch Deferred on a separate thread via writing launch() instead of launch(Unconfined). To be on a safe side during migration refactoring I am introducing thread assertions to throw an exception if multithreading happens. Also I am using require() method instead of assert(), because assert() is not actually throwing an exception - assertions are disabled by default in Kotlin.

Test: manually tested app will crash if we write just launch() for example in SchedulerBindingImpl.scheduleWarmUpFrame(), unit and integration tests are still passing.
Change-Id: I89a7fa8c239b441c3443e71d4da2fac8ecb370b0