Fix font not applied to toggle and switch buttons
The bug causes the text for toggle buttons and switch buttons to not use the font specified in the fontFamily attribute for older API levels (pre-26). As AppCompatTextHelper handles this situation, added a member of type AppCompatTextHelper for class SwitchCompat and also created an AppCompatToggleButton class with the same behaviour.

Bug: 70541754
Test: ./gradlew appcompat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.appcompat.widget.AppCompatToggleButtonTest\#testFontResources
Test: ./gradlew appcompat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.appcompat.widget.SwitchCompatTest\#testFontResources
Change-Id: I7e375f861dcbd4568034ef16d0481ae92486c0b1