Detect <include> graphs as an included destination and not a nested one.
Even though included graphs can be considered nested graph, these are
not currently parsed by the plugin and end up being a stub destination.
This causes problems when there is a global action since a class will
be generated for the nested graph containing the global action, but
since the stub destination has no id nor name, the plugin fails.

This change adds a model, 'IncludedDestination' that represents an
included graph detected by the parser and is stored in a different set
within the containing Destination model. The included graph is not
further discovered. However, by treating it different than a nested
destination we avoid the issue described above. Moreover, this change
helps toward later supporting parsing included graphs.

Bug: 116542123
Test: ./gradlew \
      :navigation:navigation-safe-args-gradle-plugin:test \
      :navigation:navigation-safe-args-generator:test

Change-Id: I9eba5706e0209247a3a599f8e84d56d36b97938e