Only check for recycling on API 21+
Prior to this there is no good way to check. On some API levels the Resources is set to null, but on others it isn't. Since the implementation does not vary by API level as long as it passes on new API levels we can be confident it has the correct behavior on older ones.

Bug: 116664299
Test: gw :core-ktx:conCh
Change-Id: Ieb7a04f727b9c2a39b5e4106bf67e5804933e855