Don't use reference id on the service side for clickable span action.
Action id's need to be stable across builds.
An accessibility service and an application might be using different versions of support library, but both ends expect the action id to be the same.
So get it on the app side, and put it in extras which the service can read.

Test: Unit tests to make sure nothing was regressed.
Change-Id: Ia3f1a2a35903d966edccf6c33ec3e8d2ebf68312