Allow trailing lambdas in place of FragmentFactory
Since in most cases only the Fragment under
test is being created, creating a full FragmentFactory
is unnecessary. Instead, we can use a trailing lambda
instantiation method to call the appropriate constructor.

This has the nice side effect of not needing to say
launchFragment<MyFragment> since the type can be
inferred from the result type of the lambda.

Test: new FragmentScenarioKotlinTest
BUG: 118898240
Change-Id: I2fc4f81012dce741460b810830e8fd37954bab45